+++ Compras/models.py
@@ -14,7 +14,8 @@ class Compra(Base):
     identificadorCompra = models.IntegerField()
     fornecedor = models.ForeignKey('Compras.Fornecedor', verbose_name='Fornecedor', on_delete=models.CASCADE)
     produto = models.ForeignKey('Produtos.Produto', verbose_name='Produto', on_delete=models.CASCADE)
-    conta = models.ForeignKey('core.Conta', verbose_name='Produto', on_delete=models.CASCADE)
+    #conta = models.ForeignKey('Contas.Conta', verbose_name='Produto', on_delete=models.CASCADE)
+    conta = models.IntegerField()
     quantidadeProduto = models.FloatField(default=0)
     precoProduto = models.FloatField(default=0)
     frete = models.FloatField(default=0)
