+++ core/views.py
@@ -4,8 +4,7 @@ from django.shortcuts import render
 from django.http import HttpResponseRedirect
 from django.contrib import messages
 
-from .models import Conta, CategoriaConta, MovimentacaoConta,  \
-     Cliente, Cartao
+from .models import Cliente
 
 from Vendas.models import Venda
