+++ unicatech/settings.py
@@ -43,6 +43,7 @@ INSTALLED_APPS = [
     'Vendas',
     'Compras',
     'Produtos',
+    'Contas',
 ]
 
 MIDDLEWARE = [
